pyproject.toml
[build-system]
equires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# 定义项目元数据
name = "yyxx-game-pkg"
version = "0.0.0"
description = "yyxx game custom module"
authors = ["yyxx-game"]
license = "MIT"
homepage = "https://github.com/yyxxgame/yyxxgame-pkg"
repository = "https://github.com/yyxxgame/yyxxgame-pkg"
readme = "README.md"
packages = [
  { include = "yyxx_game_pkg"}, {include = "tests"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.7, <4"
toml = "^0.10.2"
opentelemetry-api = "^1.16.0"
opentelemetry-exporter-jaeger = "^1.16.0"
opentelemetry-sdk = "^1.16.0"
opentelemetry-instrumentation-requests = "^0.37b0"
requests = "^2.28.2"
line-profiler = "^4.0.3"
celery = "^5.2.7"
fastapi = "^0.94.0"
uvicorn = "^0.21.0"
redis = "^4.5.1"
opentelemetry-instrumentation-fastapi = "0.37b.0"
pika = "^1.3.1"
